serialize: Use buffered reads/writes
A not obvious trap in Rust is that e.g. `File` is just a raw wrapper
for a file descriptor, and e.g. serde will happily try to do things
like invoke `read()/write()` a byte or 4 bytes at a time.

Buffering is important to avoid lots of syscall traffic, particularly
in the age of Spectre etc.

Signed-off-by: Colin Walters <walters@verbum.org>