Use anyhow for errors, drop serialize wrappers
In my opinion, creating custom error types is a "heavyweight"
operation that one should use only when:

- It is likely that a user of your library would want to *take specific
  action* on an error type.  The classic example here being
  "catch ENOENT" as separate from all othe file errors
- The other case is where one is writing a truly "foundational" library
  like a libc wrapper, or serde and you can't easily predict all use
  cases.

In this case, to my understanding this code is just going to be
executed as a subprocess in the near future.  We don't even necessarily
need to publish to crates.io.

Hence, using `anyhow` for error handling simplifies everything.

We also can just drop the custom serialization wrapper code; it doesn't
add much over just directly invoking `serde_json`.

Signed-off-by: Colin Walters <walters@verbum.org>