Initial implementation of firewalld backend
Presently, we use one zone for all interfaces, with one policy
for that zone to enable masquerading. Every time a zone or policy
is added, we need to do a firewall reload, so we can only make
one zone of each (the very first time netavark runs) unless we
want to wipe out our port-forwarding rules every time we run.
This differs a lot from my original architecture, but I think
it's the best we can do given the constaints.

Signed-off-by: Matthew Heon <matthew.heon@pm.me>