Fix double json in error message
If the NetavarkError is wrapped twice so you get json in the actual error
message field. To fix this we need to keep the default string display
and use an explicit print_json function.

This is the current message `{"error":"{\"error\":\"No such file or directory (os error 2)\"}"}`
Now it looks like this `{"error":"No such file or directory (os error 2)"}`

Signed-off-by: Paul Holzinger <pholzing@redhat.com>