Change to making only a single call to setup/tear down
This is necessary to simplify operation for port forwarding via
firewalld. Instead of requiring two separate calls to add and
remove port forward rules for v4 and v6, both are now carried in
the struct, which means I don't have to worry about one of my
DBus calls messing up the other. For iptables it's mostly a wash,
the code gets a slight bit more complex, but not significantly.

Signed-off-by: Matthew Heon <matthew.heon@pm.me>