dns: double fork aardvark server when using scoped systemd-run
It seems when running with `systemd-run` using --scope process still lingers
around as a direct child as a result script based solution like `bats` forever
waits for aardvark to end. Which is not supposed to happen since aardvark is a long
running process. As a workaround we will do double-fork for such cases.

Signed-off-by: Aditya R <arajan@redhat.com>