Update rtnetlink dependency
Update to the latest version instead of using main.
There are some breaking changes which needed adjustment,
set_name_filter() was moved to match_name(). This changed the behaviour
when the interface is not found, it now returns ENODEV so we have to
handle this instead of OK(None).
The good thing about this is that we only query a single interface.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>