test/helpers.bash: assert handle corner cases correctly
Right now `assert t:[1] != t:[1]` passes. This is obviously incorrect
and some test might not be working correctly because of this. We have to
special case this like the "==" case.

I discovered this in buildah, since we use the same logic here we should
fix it to prevent unexpected bugs.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>