Add a dedicated NetavarkError type
Subsequently, plumb this error through most of Netavark. The
interface creation logic is still mostly on std::io::Error, but
that natively converts into NetavarkError via From so I'm not
touching that right now.

Still to do: error equality, and possibly dedicated types for
iptables and firewalld so we can better track where in the code
errors occur.

Also add a NetavarkResult and convert most of our code to use it.

Signed-off-by: Matthew Heon <matthew.heon@pm.me>