setup/teardown: also go through port forwarding when no mapping
Use an empty vec if port_bindings is None, so we can always go through
port binding setup and teardown as long as an IP is set.

This is a requirement to setup port forwarding for DNS requests when
binding aardvark-dns to an ephemeral port as that defines its own port
forwarding.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>