Cirrus: Gather arm and aarch64 netavark binaries
In order to test `aardvark-dns` (and maybe elsewhere) a recent binary
build of netavark from the CI system is needed.  Previously this was
always an `x86_64` binary along with a `.debug` version.  Add support for
archiving both `arm` (32 bit) and `aarch64` binaries.  Name them with
the architecture-target as a suffix so they may be distinguished from
the (x86_64) `netavark` binary.
e.g. `bin/netavark.aarch64-unknown-linux-gnu`

Signed-off-by: Chris Evich <cevich@redhat.com>