Cirrus: Speed up CI re-runs and pushes to same PR
During development, it's common to force-push small changes or re-run
possible flakes in CI.  This is sped up significantly for rust, by
caching downloaded dependencies and compiler output.  However,
previously this cache was limited to only the same Cirrus build UUID
number.  Improve this by scoping the cache to the PR number, branch, or
tag name.

Also include the host architecture in the key so a future commit can
distinguish them.

Signed-off-by: Chris Evich <cevich@redhat.com>