fixup isolate networks test
Fix some problems with the isolation test. Fix the overlapping subnets
to make sure we correctly route. Add `-w 1` to ping to prevent it from
just hanging. Turn off dns on the network since we do not need it and
fix the test comments to actually reflect what we do.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>