aarvdark delete_entry: fix removal of empty file on dual stack
when both ipv4 and ipv6 are present we cannot just check file length
before deletion as a criteria: count the amount of lines we really
wrote instead.
Also fix the index counting for newline as well, that was only meant
to catch the final empty line caused by the split -- do this by using
split_terminator() instead.

That fixes aardvark not shutting down after all countainers are down
in dual stacked network.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>