create network driver interface
I am sorry, I know that this is almost to review properly but I couldn't
figure out how to split this out in smaller pieces.

We should abstract the internal driver logic behind a interface. This
makes to code easier to maintain since each driver is isolated.

Also this will allow us much better error handling, in case of errors it
is important to still clean up as much as possible, i.e. when we setup
two networks and the second one fails we should teardown the first as
well.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>