Cirrus: Verify expected binary artifacts present
Other CI systems depend/expect a full set of compiled binaries are
included in the success task's binary artifact (zip file).  However,
it's possible some mistake in CI could silently lead to one or more
binaries not being produced.  For example, a `bin_cache` failure.  Fix
this by verifying all expected/required files are present.  If not,
print the name of the missing binary to assist in
troubleshooting/debugging.

Signed-off-by: Chris Evich <cevich@redhat.com>