Do not use ipv6 autoconf
When we run container netavark (or in the future the dhcp proxy) should
set ip addresses. The ipv6 auto configuration is not wanted in almost
all cases so we just disable it for now.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2126243

I was able to reproduce this and can confirm that this fixes the issue
for me. I don't think we can test it in CI.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>