netlink: add del_{addr,route} calls
It is not required for netavark since we just delete links on teardown
but the dhcp proxy might need to remove addresses and/or routes when
it gets a new lease.

Unfortunately the del_route unit test doe snot seem to work since it
requires "real" working routes which we cannot create in a custom netns.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>