Add Lintje commit validation to Cirrus config
This text does not describe the change at all.  However, the dumb bot
doesn't know the difference because bots can't read!  Really, this
message is only taking up some space, to confirm the new check will
not flag this commit as "bad".  However, this commit does NOT have a
DCO, and so the bot _should_ flag that problem.