netlink: improve error message on invalid netlink result
While unlikely when netlink returns something unexpected we should
provide more details so we know where it failed. This will most likely
only fail when we called something incorrectly or due a kernel bug.
If it run with trace log level we already display the raw netlink
messages so that should be enough info to start debugging.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>