macvlan: fix name collision on hostns
When the container interface and host interface use the same link
netavark fails with EEXISTS. This is a weird kernel behavior because we
directly create the interface in the netns but the kernel still checks
in the host ns and thus it will fail if the same name is already used
there.

To fix this we need to create the interface with a tmp name then rename
it in the netns. For performance reason we still try the quick way first
and only fallback to tmp name if required.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>