Add support for route metrics
Metrics allow us to specify which default route will be preferred
when a container connects to multiple networks. By default,
all routes will get a metric of 100. If two routes go to the same
location (e.g. multiple default routes, one for each network, as
Netavark does by default), the kernel by default round-robins
traffic between all of them, but if the metrics for the routes
differ, the route(s) with the lowest metric are chosen.

For confusing reasons, metric is actually called preference in
Netlink - and, doubly confusing, there is also a "metrics" option
for routes in Netlink as well. Which, naturally, is completely
unrelated to route metric. Sigh.

Signed-off-by: Matthew Heon <matthew.heon@pm.me>