remove number of .clone() calls
In many places we really do not need to clone the values and can just
use references with lifetimes.
I also removed the mut declaration in the aardvark interface which is not
needed and thus just causes confusion.

There is a lot more I could do but these ones were very simple to fix
and I don't want to spend more time on this right now.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>