Cirrus: Improve caching effectiveness
Prior to this commit, the cache maintained for the main branch has grown
to over 16gigabytes.  Attempts to upload it are causing frequent branch-
level CI failures due to being larger than the maximum allowable.
Attempt to fix this by introducing a cache-grooming script and altering
the cache-keys to track with dependency specification and build settings.
Content for the grooming script comes mainly from three places:

1. [Cargo documentation](https://doc.rust-lang.org/nightly/cargo/guide/build-cache.html)
2. The rust-cache github action [(typescript) source code](https://github.com/Swatinem/rust-cache/tree/master/src)
3. This [cargo home in CI section](https://doc.rust-lang.org/nightly/cargo/guide/cargo-home.html#caching-the-cargo-home-in-ci)

For branch-level cache, the grooming script will completely clobber it's
cache at loosely-random intervals determined by the Cirrus Build ID.
This is intended to prevent the cache from growing indefinitely.  It's an
non-ideal solution in a space full of very limited possibilities.

Signed-off-by: Chris Evich <cevich@redhat.com>