combine dhcp-proxy and netavark into one binary
The proxy will now be called as `netavark dhcp-proxy ...`. The new
binary is about 18 MB big, this safes about 10 MB in space compared to
both binaries together.

Note: This also removes one of the duplciated server.rs files which was
in both src/dhcp_proxy and src/dhcp_proxy_server. Instead move it to the
commands dir to follow the common code structure.

Fixes #608

Signed-off-by: Paul Holzinger <pholzing@redhat.com>