iptables: port fw handle wildcard address correctly
When using `0.0.0.0` or `::` as host ip in the port spec we currently
generate incorrect rules. A wildcard address means that we should
forward traffic for all ips, e.g. that same that we do when no host ip
is set. Instead we used it literally which caused the rule only to match
traffic with a destination of `0.0.0.0` or `::` and this is not
possible.

Also fixes a problem where an invalid host ip was treated as match all.
The correct behavior is of course to throw an error. This was not a big
deal because podman already validates the ip so you could only trigger
it by manually calling netavark.

And lastly I fixed some typos in the test names as I found them
confusing. They do not actually test a range.

Fixes #636

Signed-off-by: Paul Holzinger <pholzing@redhat.com>