dhcp-proxy: apply new ip address/gateway
If we get a new lease with different ips or a new gateway we have to
update the contianer netns with the new addresses. The added tests takes
over 2m because the minimal lease time that dnsmasq supports is 2m so we
have to love with that for now.

One outstanding issue is that podman has no idea that things changed, it
will continue to show incorrect network info in podman inspect. This
never worked with CNI either so it should be ok for now.

However I think it would be a great improvement for long running
containers if we could somehow update the satus in podman. I think we
need some hidden podman command callback where we can feed podman the
new info.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>