bump nix to 0.27.1
The nix crate removes all features by default so we need to explicitly
list what we use.
Second, the io API's changed and all use the AsFd trait now. This is
great for IO safety but will require larger changes to migrate from
RawFd. Thus I went the easy way and use unsafe (not really unsafe here)
and construct and a new BorrowedFd object (which impl AsFd) for the
setns call.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>