firewall: add state functions to serialize configs
This will be used for the firewalld reload service.
On firewall setup we write the configs to disk (tmpfs) and then each
time the firewalld reload can read the configs at a later point when it
needs to readd rules. On teardown we need to remove the config again.

This commit does not yet integrate any of this it just adds the code to
do the writing, reading or removing of the config.

Also in order to deserialize the configs we need a owned struct, thus we
cannot use the PortForwardConfig with references. Becuase I like to
avoid the unnecessary clones on the setup path I created two structs:
PortForwardConfig and PortForwardConfigOwned which should help with
that.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>