firewall-reload: integrate actual logic to reload rules
First since the service is started after firewalld we assume it cleared
the rules and thus we need to relaod rules right away.
Then listen for the firewalld reloaded signal to reload later.

Only store the rules as root as rootless has no need for the service and
we do not need to pay the overhead of the writes.

The --config options is now always required for setup, teardown and
update. We always set it in podman so it is not a problem.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>