firewall/state: make sure to ignore enoent on read
When we read the firewall configs we should assume ENOENT is no error
because the reload service might be starte before any container was set
up. Instead Return a Option<> so the caller can just skip the fw setup
in this case.

To avoid boilerplate I added some macros to deal with it better.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>