firewalld-reload: add integration tests
Add a integration test for the firewalld-reload command. This makes sure
we are actually adding the rule back and are listing for the correct
event.

Unfortunately this is racy as the services run in the background so I
have to add sleep 1 to wait for the rules to be added back. Hopefully
this is enough to avoid flakes in CI.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>