firewalld-reload: prevent race which could leak fw rules
There is a small possibility that the reload service could add fw rules
after the container was teardown and thus leak them forever.

Consider this case if both processes run in parallel:

| netavark fw reload service  |  netavark teardown |
|-----------------------------|--------------------|
| read fw config              |                    |
|                             | remove fw config   |
|                             | remove fw rules    |
| add fw rules                |                    |

In order to prevent the race, this commit adds a lock file to ensure
that it is not possible to add/remove fw configs while the reload
service is doing it's thing.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>