firewalld-reload: fix CI tests
A new change in firewalld 2.0 no longer flushes all rules. This means
the test cannot check for it. Instead we must check for the trusted
sources that are added in firewalld.

This is causing CI failures on main right now because this test was
merged with CI running on f38 while the update to f39 was just merged
before that.

https://github.com/containers/netavark/pull/826
https://github.com/containers/netavark/pull/840

Signed-off-by: Paul Holzinger <pholzing@redhat.com>