use OsString/Path over String for file paths
Rust String type has a pretty well documented difference to most string
types in many other languages. A rust string must be valid utf-8.
However linux itself does not care about the char encoding in paths, it
is simply a non zero byte array. This means that technically netavark
will not accept all valid paths. While it unlikely that anyone is using
non utf-8 paths in the real world it is still a non great to deny that.

Rust actually offers the OsString/OsStr types that do not enforce the
utf-8 requirement. And there is also PathBuf/Path types that futher
abstract file paths. So to fix this just use these types instead.

See the added test which checks for it, without this patch we would see
a `error: invalid UTF-8 was detected in one or more arguments` message.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>