aardvark: show error if process is in wrong netns
In the rootless netns case it is possible that aardvark-dns kept running
to an improper teardown (root cause unclear). Next time a new rootless
netns is created it means that aardvark-dns runs in the old wrong one
and no dns will work.

Technically it is possible for us to just kill it and start a new one
but if there are other old containers running based on that netns they
would loos dns as well. There is really no good way to fix this by
ourself as we do not know if other containers are still up and running.

So instead log a error with instructions on how to resolve the situation.

Fixes containers/podman#20396

Signed-off-by: Paul Holzinger <pholzing@redhat.com>