fix netavark update to not start a new aardvark-dns
While commit 45f158118d fixed one problem it added another one as well.
The condition of !is_update got mixed up and now never returned in the
good case thus the code went to the end of the function and started a
new aardvark-dns conflicting with an existing running arrdvark-dns.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>