nftables: only dump netavark table rules
This makes things easier as we the lib no longer needs to parse
unrelated rules, we only need to touch/care about the netavark table.

One big issue is that the rust nftables lib is not feature complete and
all rules that it doesn't understand cause the json deserialization to
fail hard and thus easily break netavark. With this fix we only dump our
rules and because we the rules via this lib we should never encounter
unsupported rules and thus it should always work.

I assume this also results in a nice speed up if there are many rules in
other tables on the host as they will never be processed by the paring
logic.

Unfortunately this implementation has one big issue, nft will fail when
the table does not exists. This is normal on the first run after boot of
course. To fix this we need to add some special error handling to read
the stderr and we get the ENOENT error message on stderr so we know when
this is the case and can ignore silently.

Fixes #942

Signed-off-by: Paul Holzinger <pholzing@redhat.com>