Add suffix to Aardvark internal network filenames
Use the %int suffix for internal network files passed to AV so we
can properly trigger DNS handling (or, more accurately, lack of
DNS handling) over there. Also, to handle version migrations, if
there is a previous file for the network named without the
suffix, delete that.

This will unblock [1].

[1] https://github.com/containers/aardvark-dns/pull/447

Signed-off-by: Matt Heon <mheon@redhat.com>