Add conditional compilation of default firewall driver
The NETAVARK_DEFAULT_FW environment variable controls the default
firewall driver that will be used by the compiled Netavark.
Currently supported values are "iptables" (the default, if the
environment variable is unset), "nftables", and "none" (we'll add
"firewalld" as a supported value once that driver is done).

Unsupported values result in a panic/failure to build.

Signed-off-by: Matt Heon <mheon@redhat.com>