fix aardvark-dns error handling
Big ops, we didn't handle any error when starting aardvark-dns. While we
waited for the fork to happen we never checked the exit status of the
command thus all errors were silently ignored. Also the command is
supposed to capture the error message so we can return it as proper
error string to podman. Before we should wrote to stderr which will not
show up for podman-remote users.

Of course right now aardvark-dns doesn't actually report many errors so
we never cought this but I am trying to make bind failures visible to
users so we need this.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>