add new rust image to check for MSRV
Add a new image based of the offical rust image so we can expliclty
check for the MSRV and not depend on distro version where we never
really know what version we are getting.

We cannot use the offical image directly as we need protobuf to be
installed so create a new Dockerfile and add a script to push it to
quay.io/libpod.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>