network: bridge: add support for unmanaged mode
While Linux doesn't support modes on bridges, we use this concept to let
the user tell us if they want podman/netavark to own the bridge or not.
Managed behaves the same way as before this commit. Unmanaged requires
the bridge to exist already, will not setup any sysctls or firewall
rules on the host and will not delete the bridge once all containers
left.

Fixes #1090

Signed-off-by: Michael Zimmermann <sigmaepsilon92@gmail.com>