nftables: add daddr match to port forward jump rule
When we have the same port forwarded with two different host ips there
was a rule ambiguity in the NETAVARK-HOSTPORT-DNAT. Each port would have
the same jump rule and teardown would have no idea when to remove this
rule easily.

To fix this just add the daddr condition to the jump rule as well. That
way we know the rule will be unique and there is only one match on
teardown.

Fixes #1129

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit 5861afa485b420590aa3ff71349a52fccf35055b)