bridge: add vlan support
Implement a new vlan option that turn the bridge vlan_filtering mode on
and then assign the host side veth to the given vlan id.

With that a users should be able to have multiple networks with the same
bridge but different vlan ids (we need to change network create logic fo
this).

Fixes #1028

Signed-off-by: Paul Holzinger <pholzing@redhat.com>