Makefile: do not rebuild if nothing changed
If the src has not changed there is no need to rebuild.
This removes the -debug suffix from the binary if build with debug set,
I doubt anyone actually uses this and we need the same name for the
bin/netavark target to actually work. This is even better because now
you could build a debug binary and run the test with it without changing
the test path manually.

Fixes #1178

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit a08bb557fd606e12cc5b6452df0e88fbc544fe67)
Signed-off-by: Paul Holzinger <pholzing@redhat.com>