Fix detect of Firewalld's StrictForwardPorts property
The code was written to work with a boolean property.
Unfortunately, it's not a boolean, it's a DBus Variant holding a
String which can be either "yes" or "no". This, unfortunately,
complicates our decoding code somewhat considerably.

I don't know how this actually worked when I tested it in the
first place, unless the API type changed in firewalld before
final release, which I don't *think* happened? Fortunately, the
code here is written to be very forgiving of errors, but I added
some warning logs to give us a trail if stuff starts blowing up
in the future (because Variant means the type can, theoretically,
change in the future, which would be unpleasant).

Fixes https://issues.redhat.com/browse/RHEL-80266

Signed-off-by: Matt Heon <mheon@redhat.com>